+++ air/src/air/tests.rs
@@ -205,7 +205,16 @@ impl MockAir {
         let mut result = Self::new(
             TraceInfo::with_meta(4, trace_length, vec![1]),
             (),
-            ProofOptions::new(32, 8, 0, FieldExtension::None, 4, 31, BatchingMethod::Linear),
+            ProofOptions::new(
+                32,
+                8,
+                0,
+                FieldExtension::None,
+                4,
+                31,
+                BatchingMethod::Linear,
+                BatchingMethod::Linear,
+            ),
         );
         result.periodic_columns = column_values;
         result
@@ -215,7 +224,16 @@ impl MockAir {
         let mut result = Self::new(
             TraceInfo::with_meta(4, trace_length, vec![assertions.len() as u8]),
             (),
-            ProofOptions::new(32, 8, 0, FieldExtension::None, 4, 31, BatchingMethod::Linear),
+            ProofOptions::new(
+                32,
+                8,
+                0,
+                FieldExtension::None,
+                4,
+                31,
+                BatchingMethod::Linear,
+                BatchingMethod::Linear,
+            ),
         );
         result.assertions = assertions;
         result
@@ -265,7 +283,16 @@ pub fn build_context<B: StarkField>(
     trace_width: usize,
     num_assertions: usize,
 ) -> AirContext<B> {
-    let options = ProofOptions::new(32, 8, 0, FieldExtension::None, 4, 31, BatchingMethod::Linear);
+    let options = ProofOptions::new(
+        32,
+        8,
+        0,
+        FieldExtension::None,
+        4,
+        31,
+        BatchingMethod::Linear,
+        BatchingMethod::Linear,
+    );
     let t_degrees = vec![TransitionConstraintDegree::new(2)];
     let trace_info = TraceInfo::new(trace_width, trace_length);
     AirContext::new(trace_info, t_degrees, num_assertions, options)