+++ air/src/proof/context.rs
@@ -18,18 +18,25 @@ pub struct Context {
     trace_info: TraceInfo,
     field_modulus_bytes: Vec<u8>,
     options: ProofOptions,
+    num_constraints: usize,
 }
 
 impl Context {
     // CONSTRUCTOR
     // --------------------------------------------------------------------------------------------
-    /// Creates a new context for a computation described by the specified field, trace info, and
-    /// proof options.
+    /// Creates a new context for a computation described by the specified field, trace info, proof
+    /// options, and total number of constraints.
     ///
     /// # Panics
-    /// Panics if either trace length or the LDE domain size implied by the trace length and the
-    /// blowup factor is greater then [u32::MAX].
-    pub fn new<B: StarkField>(trace_info: TraceInfo, options: ProofOptions) -> Self {
+    /// - If either trace length or the LDE domain size implied by the trace length and the
+    ///   blowup factor is greater then [u32::MAX].
+    /// - If the number of constraints is not greater than 0.
+    /// - If the number of constraints is greater than [u32::MAX].
+    pub fn new<B: StarkField>(
+        trace_info: TraceInfo,
+        options: ProofOptions,
+        num_constraints: usize,
+    ) -> Self {
         // TODO: return errors instead of panicking?
 
         let trace_length = trace_info.length();
@@ -38,10 +45,14 @@ impl Context {
         let lde_domain_size = trace_length * options.blowup_factor();
         assert!(lde_domain_size <= u32::MAX as usize, "LDE domain size too big");
 
+        assert!(num_constraints > 0, "number of constraints should be greater than zero");
+        assert!(num_constraints <= u32::MAX as usize, "number of constraints is too big");
+
         Context {
             trace_info,
             field_modulus_bytes: B::get_modulus_le_bytes(),
             options,
+            num_constraints,
         }
     }
 
@@ -84,6 +95,11 @@ impl Context {
     pub fn options(&self) -> &ProofOptions {
         &self.options
     }
+
+    /// Returns the total number of constraints.
+    pub fn num_constraints(&self) -> usize {
+        self.num_constraints
+    }
 }
 
 impl<E: StarkField> ToElements<E> for Context {
@@ -105,6 +121,9 @@ impl<E: StarkField> ToElements<E> for Context {
         let (m1, m2) = self.field_modulus_bytes.split_at(num_modulus_bytes / 2);
         result.push(E::from_bytes_with_padding(m1));
         result.push(E::from_bytes_with_padding(m2));
+
+        // convert the number of constraints
+        result.push(E::from(self.num_constraints as u32));
 
         // convert proof options to elements
         result.append(&mut self.options.to_elements());
@@ -148,7 +167,15 @@ impl Deserializable for Context {
         // read options
         let options = ProofOptions::read_from(source)?;
 
-        Ok(Context { trace_info, field_modulus_bytes, options })
+        // read total number of constraints
+        let num_constraints = source.read_usize()?;
+
+        Ok(Context {
+            trace_info,
+            field_modulus_bytes,
+            options,
+            num_constraints,
+        })
     }
 }
 
@@ -170,6 +197,8 @@ mod tests {
         let grinding_factor = 20;
         let blowup_factor = 8;
         let num_queries = 30;
+        let num_constraints = 128;
+        let batching_constraints = BatchingMethod::Linear;
         let batching_deep = BatchingMethod::Linear;
 
         let main_width = 20;
@@ -178,10 +207,10 @@ mod tests {
         let trace_length = 4096;
 
         let ext_fri = u32::from_le_bytes([
+            blowup_factor as u8,
             fri_remainder_max_degree,
             fri_folding_factor,
             field_extension as u8,
-            0,
         ]);
 
         let expected = {
@@ -197,9 +226,9 @@ mod tests {
             expected.extend(vec![
                 BaseElement::from(1_u32),    // lower bits of field modulus
                 BaseElement::from(u32::MAX), // upper bits of field modulus
+                BaseElement::from(num_constraints as u32),
                 BaseElement::from(ext_fri),
                 BaseElement::from(grinding_factor),
-                BaseElement::from(blowup_factor as u32),
                 BaseElement::from(num_queries as u32),
             ]);
 
@@ -213,11 +242,12 @@ mod tests {
             field_extension,
             fri_folding_factor as usize,
             fri_remainder_max_degree as usize,
+            batching_constraints,
             batching_deep,
         );
         let trace_info =
             TraceInfo::new_multi_segment(main_width, aux_width, aux_rands, trace_length, vec![]);
-        let context = Context::new::<BaseElement>(trace_info, options);
+        let context = Context::new::<BaseElement>(trace_info, options, num_constraints);
         assert_eq!(expected, context.to_elements());
     }
 }